requirements.txt
alembic==1.12.1 ; python_version >= "3.9" and python_version < "4.0"
annotated-types==0.6.0 ; python_version >= "3.9" and python_version < "4.0"
anyio==3.7.1 ; python_version >= "3.9" and python_version < "4.0"
astroid==3.0.1 ; python_version >= "3.9" and python_version < "4.0"
bcrypt==4.1.1 ; python_version >= "3.9" and python_version < "4.0"
black==23.11.0 ; python_version >= "3.9" and python_version < "4.0"
boto3==1.29.7 ; python_version >= "3.9" and python_version < "4.0"
botocore==1.32.7 ; python_version >= "3.9" and python_version < "4.0"
build==1.0.3 ; python_version >= "3.9" and python_version < "4.0"
cachecontrol[filecache]==0.13.1 ; python_version >= "3.9" and python_version < "4.0"
certifi==2023.11.17 ; python_version >= "3.9" and python_version < "4.0"
cffi==1.16.0 ; python_version >= "3.9" and python_version < "4.0" and (sys_platform == "darwin" or sys_platform == "linux")
cfgv==3.4.0 ; python_version >= "3.9" and python_version < "4.0"
charset-normalizer==3.3.2 ; python_version >= "3.9" and python_version < "4.0"
cleo==2.1.0 ; python_version >= "3.9" and python_version < "4.0"
click==8.1.7 ; python_version >= "3.9" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.9" and python_version < "4.0" and (sys_platform == "win32" or platform_system == "Windows" or os_name == "nt")
crashtest==0.4.1 ; python_version >= "3.9" and python_version < "4.0"
cryptography==41.0.5 ; python_version >= "3.9" and python_version < "4.0" and sys_platform == "linux"
dill==0.3.7 ; python_version >= "3.9" and python_version < "4.0"
distlib==0.3.7 ; python_version >= "3.9" and python_version < "4.0"
dulwich==0.21.6 ; python_version >= "3.9" and python_version < "4.0"
ecdsa==0.18.0 ; python_version >= "3.9" and python_version < "4.0"
exceptiongroup==1.2.0 ; python_version >= "3.9" and python_version < "3.11"
fastapi==0.104.1 ; python_version >= "3.9" and python_version < "4.0"
fastjsonschema==2.19.0 ; python_version >= "3.9" and python_version < "4.0"
filelock==3.13.1 ; python_version >= "3.9" and python_version < "4.0"
greenlet==3.0.1 ; python_version >= "3.9" and python_version < "4.0" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32")
h11==0.14.0 ; python_version >= "3.9" and python_version < "4.0"
identify==2.5.32 ; python_version >= "3.9" and python_version < "4.0"
idna==3.6 ; python_version >= "3.9" and python_version < "4.0"
importlib-metadata==6.8.0 ; python_version >= "3.9" and python_version < "3.12"
installer==0.7.0 ; python_version >= "3.9" and python_version < "4.0"
isort==5.12.0 ; python_version >= "3.9" and python_version < "4.0"
jaraco-classes==3.3.0 ; python_version >= "3.9" and python_version < "4.0"
jeepney==0.8.0 ; python_version >= "3.9" and python_version < "4.0" and sys_platform == "linux"
jmespath==1.0.1 ; python_version >= "3.9" and python_version < "4.0"
keyring==24.3.0 ; python_version >= "3.9" and python_version < "4.0"
load-dotenv==0.1.0 ; python_version >= "3.9" and python_version < "4.0"
mako==1.3.0 ; python_version >= "3.9" and python_version < "4.0"
markupsafe==2.1.3 ; python_version >= "3.9" and python_version < "4.0"
mccabe==0.7.0 ; python_version >= "3.9" and python_version < "4.0"
more-itertools==10.1.0 ; python_version >= "3.9" and python_version < "4.0"
msgpack==1.0.7 ; python_version >= "3.9" and python_version < "4.0"
mypy-extensions==1.0.0 ; python_version >= "3.9" and python_version < "4.0"
mysql-connector-python==8.2.0 ; python_version >= "3.9" and python_version < "4.0"
nodeenv==1.8.0 ; python_version >= "3.9" and python_version < "4.0"
packaging==23.2 ; python_version >= "3.9" and python_version < "4.0"
passlib[bcrypt]==1.7.4 ; python_version >= "3.9" and python_version < "4.0"
pathspec==0.11.2 ; python_version >= "3.9" and python_version < "4.0"
pexpect==4.9.0 ; python_version >= "3.9" and python_version < "4.0"
pkginfo==1.9.6 ; python_version >= "3.9" and python_version < "4.0"
platformdirs==3.11.0 ; python_version >= "3.9" and python_version < "4.0"
poetry-core==1.8.1 ; python_version >= "3.9" and python_version < "4.0"
poetry-plugin-export==1.6.0 ; python_version >= "3.9" and python_version < "4.0"
poetry==1.7.1 ; python_version >= "3.9" and python_version < "4.0"
pre-commit==3.5.0 ; python_version >= "3.9" and python_version < "4.0"
protobuf==4.21.12 ; python_version >= "3.9" and python_version < "4.0"
ptyprocess==0.7.0 ; python_version >= "3.9" and python_version < "4.0"
pyasn1==0.5.1 ; python_version >= "3.9" and python_version < "4.0"
pycparser==2.21 ; python_version >= "3.9" and python_version < "4.0" and (sys_platform == "darwin" or sys_platform == "linux")
pydantic-core==2.14.5 ; python_version >= "3.9" and python_version < "4.0"
pydantic==2.5.2 ; python_version >= "3.9" and python_version < "4.0"
pylint==3.0.2 ; python_version >= "3.9" and python_version < "4.0"
pyproject-hooks==1.0.0 ; python_version >= "3.9" and python_version < "4.0"
python-dateutil==2.8.2 ; python_version >= "3.9" and python_version < "4.0"
python-dotenv==1.0.0 ; python_version >= "3.9" and python_version < "4.0"
python-jose==3.3.0 ; python_version >= "3.9" and python_version < "4.0"
python-multipart==0.0.6 ; python_version >= "3.9" and python_version < "4.0"
pywin32-ctypes==0.2.2 ; python_version >= "3.9" and python_version < "4.0" and sys_platform == "win32"
pyyaml==6.0.1 ; python_version >= "3.9" and python_version < "4.0"
rapidfuzz==3.5.2 ; python_version >= "3.9" and python_version < "4.0"
requests-toolbelt==1.0.0 ; python_version >= "3.9" and python_version < "4.0"
requests==2.31.0 ; python_version >= "3.9" and python_version < "4.0"
rsa==4.9 ; python_version >= "3.9" and python_version < "4"
s3transfer==0.8.0 ; python_version >= "3.9" and python_version < "4.0"
secretstorage==3.3.3 ; python_version >= "3.9" and python_version < "4.0" and sys_platform == "linux"
setuptools==69.0.2 ; python_version >= "3.9" and python_version < "4.0"
shellingham==1.5.4 ; python_version >= "3.9" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.9" and python_version < "4.0"
sniffio==1.3.0 ; python_version >= "3.9" and python_version < "4.0"
sqlalchemy==2.0.23 ; python_version >= "3.9" and python_version < "4.0"
starlette==0.27.0 ; python_version >= "3.9" and python_version < "4.0"
tomli==2.0.1 ; python_version >= "3.9" and python_version < "3.11"
tomlkit==0.12.3 ; python_version >= "3.9" and python_version < "4.0"
trove-classifiers==2023.11.22 ; python_version >= "3.9" and python_version < "4.0"
typing-extensions==4.8.0 ; python_version >= "3.9" and python_version < "4.0"
urllib3==1.26.18 ; python_version >= "3.9" and python_version < "4.0"
uvicorn==0.24.0.post1 ; python_version >= "3.9" and python_version < "4.0"
virtualenv==20.24.7 ; python_version >= "3.9" and python_version < "4.0"
xattr==0.10.1 ; python_version >= "3.9" and python_version < "4.0" and sys_platform == "darwin"
zipp==3.17.0 ; python_version >= "3.9" and python_version < "3.12"
